Remove legacy python MC code
Clean dead code from the `monte_carlo` module,
