+++ SymbolicDSGE/monte_carlo/serialize.py
@@ -20,7 +20,7 @@ import numpy as np
 from numpy.typing import NDArray
 
 from ..regression.ols import OLSResult
-from .mc_constructs import MCContext, MCPipelineResult
+from .mc_constructs import MCPipelineResult
 from .postproc import Artifact, Raw, Summary, normalize_artifacts
 from .traces import regression_trace_keys, test_trace_keys
 
@@ -102,7 +102,6 @@ def serialize_pipeline_result(
             name: _serialize_regression_summary(summary)
             for name, summary in result.regression_summaries.items()
         },
-        "data_summaries": _summarize_context_data(result.contexts or ()),
         "postproc": {
             name: _serialize_artifact(artifact)
             for name, artifact in _normalized_postproc(result).items()
@@ -470,63 +469,6 @@ def _status_counts(status_trace: Sequence[Any]) -> dict[str, int]:
     for status in status_trace:
         counts[status.name] = counts.get(status.name, 0) + 1
     return counts
-
-
-def _summarize_context_data(contexts: Sequence[MCContext]) -> dict[str, Any]:
-    arrays: dict[str, list[np.ndarray]] = {}
-    for context in contexts:
-        if context.data is None:
-            continue
-        data = context.data
-        if data.states is not None:
-            arrays.setdefault("states", []).append(np.asarray(data.states))
-        if data.observables is not None:
-            arrays.setdefault("observables", []).append(np.asarray(data.observables))
-        for name, value in data.raw.items():
-            if name != "_X":
-                arrays.setdefault(f"raw:{name}", []).append(np.asarray(value))
-    return {name: _array_collection_summary(values) for name, values in arrays.items()}
-
-
-def _array_collection_summary(values: Sequence[np.ndarray]) -> dict[str, Any]:
-    n_total = sum(int(arr.size) for arr in values)
-    n_finite = 0
-    value_sum = 0.0
-    square_sum = 0.0
-    minimum = np.inf
-    maximum = -np.inf
-    for arr in values:
-        finite = arr[np.isfinite(arr)]
-        if finite.size == 0:
-            continue
-        n_finite += int(finite.size)
-        value_sum += float(finite.sum())
-        square_sum += float(np.square(finite).sum())
-        minimum = min(minimum, float(finite.min()))
-        maximum = max(maximum, float(finite.max()))
-    if n_finite == 0:
-        return {
-            "n_rep": len(values),
-            "shape": list(values[0].shape),
-            "n_values": n_total,
-            "n_finite": 0,
-            "mean": None,
-            "std": None,
-            "min": None,
-            "max": None,
-        }
-    mean = value_sum / n_finite
-    variance = max(0.0, square_sum / n_finite - mean**2)
-    return {
-        "n_rep": len(values),
-        "shape": list(values[0].shape),
-        "n_values": n_total,
-        "n_finite": n_finite,
-        "mean": _json_float(mean),
-        "std": _json_float(variance**0.5),
-        "min": _json_float(minimum),
-        "max": _json_float(maximum),
-    }
 
 
 def _trace_summary(values: Any) -> dict[str, Any]:
